backend(no use)/requirements.txt
dj_database_url==3.0.1
Django==4.2.11
djangorestframework==3.16.1
djangorestframework_simplejwt==4.8.0
exponent_server_sdk==2.2.0
Pillow==10.2.0
protobuf==5.29.5
python-dotenv==1.0.0
requests==2.32.5
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
django-cors-headers==4.3.1
djoser==2.1.0
cloudinary==1.44.1
dj3-cloudinary-storage==0.0.6
google-auth==2.40.3
